backend-setup/requirements.txt
# LiveKit (using latest stable versions)
livekit-agents>=0.15.0
livekit-api>=0.8.0
livekit-plugins-deepgram
livekit-plugins-cartesia
livekit-plugins-silero

# HTTP client for HF inference
httpx>=0.24.1

# Flask API
flask>=3.0.3
flask-cors>=4.0.0
flask-socketio>=5.3.6
python-socketio>=5.9.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Authentication
pyjwt>=2.8.1
bcrypt>=4.1.1

# Configuration
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

# Utilities
aiofiles>=24.1.0
requests>=2.31.0

# Testing
pytest>=7.4.3
hypothesis>=6.92.1

# AI / ML for Semantic Scoring
sentence-transformers>=2.2.2
numpy>=1.26.0
scikit-learn>=1.3.0
